api/requirements.txt
flask         # Necessário para a API (somente na pasta api)
psycopg2-binary # Para conectar ao PostgreSQL
pika          # Para conectar ao RabbitMQ


